move relaxometry params to data folder
